+++ lib/src/utils/console_logger.dart
@@ -40,7 +40,8 @@ class ConsoleLogger {
     Logger.root.level = Level.ALL;
     Logger.root.onRecord.listen((record) {
       final prefix = _getPrefixForLevel(record.level);
-      final time = '${record.time.hour.toString().padLeft(2, '0')}:${record.time.minute.toString().padLeft(2, '0')}:${record.time.second.toString().padLeft(2, '0')}';
+      final time =
+          '${record.time.hour.toString().padLeft(2, '0')}:${record.time.minute.toString().padLeft(2, '0')}:${record.time.second.toString().padLeft(2, '0')}';
 
       // Using print here is acceptable for CLI logging output
       // ignore: avoid_print