+++ test/utils/sound_notification_test.dart
@@ -4,7 +4,10 @@ import 'package:share_my_apk/src/utils/sound_notification_util.dart';
 void main() {
   group('SoundNotificationUtil', () {
     test('playNotificationSound does not throw', () {
-      expect(() => SoundNotificationUtil.playNotificationSound(), returnsNormally);
+      expect(
+        () => SoundNotificationUtil.playNotificationSound(),
+        returnsNormally,
+      );
     });
   });
 }